Move files to directly under lib folder
