+++ lib/authentication.js
@@ -2,8 +2,8 @@
 const url = require('url')
 
 // Local dependencies
-const config = require('../../config')
-const { encryptPassword } = require('../../utils')
+const config = require('./config')
+const { encryptPassword } = require('./utils')
 
 // Local variables
 const allowedPathsWhenUnauthenticated = [
