+++ lib/authentication.test.js
@@ -4,7 +4,7 @@ const url = require('url')
 
 // Local dependencies
 const authentication = require('./authentication')
-const config = require('../../config')
+const config = require('./config')
 
 // Local variables
 const userPassword = 'secure-password'
