+++ lib/build.js
@@ -7,12 +7,12 @@ const fse = require('fs-extra')
 const nodemon = require('nodemon')
 const sass = require('sass')
 
-const plugins = require('../plugins/plugins')
+const plugins = require('./plugins/plugins')
 
-const utils = require('../utils')
+const utils = require('./utils')
 
-const buildConfig = require('./config.json')
-const { appDir, projectDir, packageDir, tempDir } = require('../utils/paths')
+const buildConfig = require('./build-config.json')
+const { appDir, projectDir, packageDir, tempDir } = require('./utils/paths')
 const { paths } = buildConfig
 
 const appCssPath = path.join(projectDir, paths.public, 'stylesheets')
