+++ lib/plugins/plugins.js
@@ -44,7 +44,7 @@ const { projectDir } = require('../utils/paths')
 const pkgPath = path.join(projectDir, 'package.json')
 const chokidar = require('chokidar')
 const fse = require('fs-extra')
-const { shadowNunjucks } = require('../build/config.json').paths
+const { shadowNunjucks } = require('../build-config.json').paths
 
 // Generic utilities
 const removeDuplicates = arr => [...new Set(arr)]
