+++ lib/session.js
@@ -4,9 +4,9 @@ const session = require('express-session')
 const FileStore = require('session-file-store')(session)
 const { get: getKeypath } = require('lodash')
 
-const { getConfig } = require('../config')
-const { projectDir, tempDir } = require('./paths')
-const { sessionFileStoreQuietLogFn } = require('./index')
+const { getConfig } = require('./config')
+const { projectDir, tempDir } = require('./utils/paths')
+const { sessionFileStoreQuietLogFn } = require('./utils')
 
 // Add Nunjucks function called 'checked' to populate radios and checkboxes
 const addCheckedFunction = function (env) {
