+++ lib/session.test.js
